moved fun from task_utils to utils
