+++ viadot/tasks/genesys.py
@@ -13,6 +13,7 @@ from prefect.utilities.tasks import defaults_from_attrs
 
 from viadot.exceptions import APIError
 from viadot.sources import Genesys
+from viadot.utils import check_value
 from viadot.task_utils import *
 
 logger = logging.get_logger()
